+++ src/main/java/com/cjm721/overloaded/client/render/entity/RenderMultiLeggings.java
@@ -1,15 +1,12 @@
 package com.cjm721.overloaded.client.render.entity;
 
-import net.minecraft.client.model.ModelBiped;
 import net.minecraft.client.renderer.GlStateManager;
 import net.minecraft.entity.Entity;
-import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.util.ResourceLocation;
 
 import javax.annotation.Nullable;
 
 import static com.cjm721.overloaded.Overloaded.MODID;
-import static com.cjm721.overloaded.Overloaded.cachedConfig;
 
 public class RenderMultiLeggings extends AbstractRenderMultiArmor {
 
@@ -38,22 +35,18 @@ public class RenderMultiLeggings extends AbstractRenderMultiArmor {
     }
 
     @Override
-    public void render(@Nullable Entity entity, float f, float f1, float f2, float f3, float f4, float f5) {
-
-        if (entity != null && entity instanceof EntityPlayer) {
-            super.setRotationAngles(f, f1, f2, f3, f4, f5, entity);
-        }
+    public void render(@Nullable Entity entity, float limbSwing, float limbSwingAmount, float ageInTicks, float netHeadYaw, float headPitch, float scale) {
+        super.setRotationAngles(limbSwing, limbSwingAmount, ageInTicks, netHeadYaw, headPitch, scale, entity);
 
         GlStateManager.pushMatrix();
-        if (entity != null && entity.isSneaking()) {
+        if (this.isSneak) {
             GlStateManager.translate(0.0F, 0.2F, 0.0F);
         }
 
-        this.bipedBody.render(f5);
-        this.bipedRightLeg.render(f5);
-        this.bipedLeftLeg.render(f5);
+        this.bipedBody.render(scale);
+        this.bipedRightLeg.render(scale);
+        this.bipedLeftLeg.render(scale);
 
         GlStateManager.popMatrix();
-
     }
 }